+++ tests/StrTest.php
@@ -8,7 +8,7 @@ use PHPUnit\Framework\TestCase;
 class StrTest extends TestCase
 {
     /** @test */
-    function it_can_humanize_string()
+    public function it_can_humanize_string()
     {
         $expected = 'Foobar Is Awesome';
         $output = Str::humanize('foobar-is-awesome');
@@ -17,7 +17,7 @@ class StrTest extends TestCase
     }
 
     /** @test */
-    function it_can_replace_string()
+    public function it_can_replace_string()
     {
         $expected = 'Orchestra Platform is awesome';
         $output = Str::replace('{name} is awesome', ['name' => 'Orchestra Platform']);
@@ -39,7 +39,7 @@ class StrTest extends TestCase
     }
 
     /** @test */
-    function it_can_convert_to_searchable_string()
+    public function it_can_convert_to_searchable_string()
     {
         $expected = ['foobar%'];
         $output = Str::searchable('foobar*');
@@ -53,11 +53,11 @@ class StrTest extends TestCase
     }
 
     /** @test */
-    function it_can_convert_stream_get_contents_string()
+    public function it_can_convert_stream_get_contents_string()
     {
-        $base_path = __DIR__.'/stub/';
+        $basePath = __DIR__.'/Stubs/';
         $expected = 'a:2:{s:4:"name";s:9:"Orchestra";s:5:"theme";a:2:{s:7:"backend";s:7:"default";s:8:"frontend";s:7:"default";}}';
-        $stream = fopen($base_path.'driver1.stub.php', 'r');
+        $stream = fopen($basePath.'driver1.stub', 'r');
         $output = Str::streamGetContents($stream);
 
         $this->assertEquals($expected, $output);
@@ -73,7 +73,7 @@ class StrTest extends TestCase
         $this->assertEquals($expected, unserialize($output));
 
         $expected = 'a:2:{s:4:"name";s:9:"Orchestra";s:5:"theme";a:2:{s:7:"backend";s:7:"default";s:8:"frontend";s:7:"default";}}'."\n";
-        $stream = fopen($base_path.'driver2.stub.php', 'r');
+        $stream = fopen($basePath.'driver2.stub', 'r');
         $output = Str::streamGetContents($stream);
 
         $this->assertEquals($expected, $output);
@@ -91,7 +91,7 @@ class StrTest extends TestCase
     }
 
     /** @test */
-    function it_can_convert_utf8_string_to_title_case()
+    public function it_can_convert_utf8_string_to_title_case()
     {
         $this->assertEquals('Taylor', Str::title('taylor'));
         $this->assertEquals('Άχιστη', Str::title('άχιστη'));